Save Independent House Functionality Added

+++ controllers/wishlistController.js
@@ -1,5 +1,6 @@
 const Wishlist = require('../models/wishlistsModel');
 const Building = require('../models/buildingsModel');
+const ResidentialHouse = require('../models/residentialHouseModel');
 const House = require('../models/housesModel');
 const catchAsync = require('../utils/catchAsync');
 const AppError = require('../utils/app-error');
@@ -31,7 +32,14 @@ exports.getMyWishlists = catchAsync(async (req, res, next) => {
 
 exports.postWishlist = catchAsync(async (req, res, next) => {
     if(req.params) {
-        req.body.building = req.params.building;
+        const building = await Building.findById(req.params.building);
+        const independentHouse = await ResidentialHouse.findById(req.params.building);
+        if(building) {
+            req.body.building = req.params.building;
+        }
+        if(independentHouse) {
+            req.body.independentHouse = req.params.building;
+        }
     }
     req.body.user = req.user;
 
@@ -46,7 +54,15 @@ exports.postWishlist = catchAsync(async (req, res, next) => {
 
 exports.deleteWishlist = catchAsync(async (req, res, next) => {
     const buildingId = req.params.building;
-    const wishlist = await Wishlist.findOneAndDelete({building: buildingId});
+    const building = await Building.findById(buildingId);
+    const independentHouse = await ResidentialHouse.findById(buildingId);
+    let wishlist;
+    if(building) {
+        wishlist = await Wishlist.findOneAndDelete({building: buildingId});
+    }
+    if(independentHouse) {
+        wishlist = await Wishlist.findOneAndDelete({independentHouse: buildingId});
+    }
 
     if(!wishlist) {
         return next(new AppError(`No documents found with that ID`, 404));
